increase capacity for event queue in Spark listener bus
In order to prevent the following error:

=======
ERROR LiveListenerBus:70 - Dropping SparkListenerEvent because no remaining room in event queue.
This likely means one of the SparkListeners is too slow and cannot keep up with the rate at which tasks are being started by the scheduler.
=======

Increasing this value may result in the driver using more memory.